Made randomGame it's own component
Didn't like how long gameList was getting so I split the randomGame parts into it's own component.
