+++ src/static/components/gameList.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import Game from './game';
 import GameSearch from './gameSearch';
-import { toast, ToastContainer } from 'react-toastify';
-import 'react-toastify/dist/ReactToastify.css';
+import RandomGame from './randomGame';
 
 export default function GameList({
 	gameList,
@@ -22,17 +21,6 @@ export default function GameList({
 		deleteGameList(gameListToDelete);
 	};
 
-	const getRandomGame = () => {
-		if (games.length > 1) {
-			const randomNum = Math.floor(Math.random() * games.length);
-			toast(`Play ${games[randomNum].igdbGame.name}`);
-		} else if (games.length == 1) {
-			toast.warn('You need to add at least one more game to this list!');
-		} else {
-			toast.warn('Add a couple games to your list first!');
-		}
-	};
-
 	return (
 		<div className="col-xs-12 col-sm-6 col-md-4">
 			<div className="p-3 border bg-light">
@@ -46,25 +34,7 @@ export default function GameList({
 					<span className="visually-hidden">Delete List</span>
 					&times;
 				</button>
-				<button
-					type="button"
-					className="btn btn-secondary btn-sm"
-					onClick={getRandomGame}
-				>
-					Don't know what to play?
-				</button>
-				<ToastContainer
-					position="top-right"
-					autoClose={5000}
-					hideProgressBar={false}
-					newestOnTop={false}
-					closeOnClick
-					rtl={false}
-					pauseOnFocusLoss
-					draggable
-					pauseOnHover
-					theme="dark"
-				/>
+				<RandomGame games={games} />
 				<GameSearch gameList={gameList} setGameList={setGameList} />
 
 				{games.map((game) => {
